refactor(settings): extract filename-template JSX sections into sibling components
Completes the M-1 split: MediaServerExportSettings and
FilenameBatchRenameSection own their job state, polling
(useSettingsJobPolling), errors, and confirm dialogs;
FilenameTemplateReference renders the catalog reference. The parent
keeps only the preset/template form and live preview and passes the
form state down (settings/onChange, namingMode, effectiveTemplate,
currentTemplateInvalid, recommendedTvLayout).

FilenameTemplateSettings.tsx: 729 -> 360 lines (1,020 before the first
pass). Typecheck clean, ESLint 0 warnings, 181 Settings tests pass.

Report: reports/codebase-review-2026-07-01.md (M-1 done)
